Change the type of the system variables innodb_segment_fill_factor and innodb_index_fill_factor to double.
These variables are used to represent percentages but could not be set
to percentages with fractional parts. Furthermore, their initial default
values were fractional percentages, but had to be changed because the
interface did not support them. Now these system variables are made of
type double and their default values are restored to their original
form.

+++ storage/innobase/include/srv0srv.h
@@ -132,8 +132,8 @@ extern char**	srv_log_group_home_dirs;
 #ifndef UNIV_HOTBACKUP
 extern ulong	srv_auto_extend_increment;
 
-extern ulong	srv_segment_fill_factor;
-extern uint	srv_index_fill_factor;
+extern double	srv_segment_fill_factor;
+extern double	srv_index_fill_factor;
 extern my_bool	srv_lease_fragment_extents;
 
 extern ibool	srv_created_new_raw;
